Deprecate BaseBlockService and BlockAdminServiceInterface
