+++ Block/BlockAdminServiceInterface.php
@@ -11,35 +11,17 @@
 
 namespace Sonata\BlockBundle\Block;
 
-use Sonata\AdminBundle\Form\FormMapper;
-use Sonata\BlockBundle\Model\BlockInterface;
-use Sonata\CoreBundle\Model\MetadataInterface;
-use Sonata\CoreBundle\Validator\ErrorElement;
+use Sonata\BlockBundle\Block\Service\AdminBlockServiceInterface;
 
-interface BlockAdminServiceInterface
+@trigger_error(
+    'This class is deprecated since 3.x and will be removed with the 4.0 release.'.
+    'Use '.__NAMESPACE__.'\Block\Service\AdminBlockServiceInterface instead.',
+    E_USER_DEPRECATED
+);
+
+/**
+ * @deprecated since 3.x, to be removed with 4.0
+ */
+interface BlockAdminServiceInterface extends AdminBlockServiceInterface
 {
-    /**
-     * @param FormMapper     $form
-     * @param BlockInterface $block
-     */
-    public function buildEditForm(FormMapper $form, BlockInterface $block);
-
-    /**
-     * @param FormMapper     $form
-     * @param BlockInterface $block
-     */
-    public function buildCreateForm(FormMapper $form, BlockInterface $block);
-
-    /**
-     * @param ErrorElement   $errorElement
-     * @param BlockInterface $block
-     */
-    public function validateBlock(ErrorElement $errorElement, BlockInterface $block);
-
-    /**
-     * @param string|null $code
-     *
-     * @return MetadataInterface
-     */
-    public function getBlockMetadata($code = null);
 }
